==> notebooks/1_exploratory_data_analysis.ipynb <==
Importar librerías necesarias

import pandas as pd

df_airbnb = pd.read_csv('/Users/melaniealvarez/Documents/Octavo semestres/Data Mining/Pset2/data/raw/Airbnb_Data.csv')
print(df_airbnb.head())

df_airbnb.shape

df_airbnb.columns

df_airbnb.info()

POsibles variables a considerar para el modelo

 0   id                      74111 non-null  int64  
 1   log_price               74111 non-null  float64
 2   property_type           74111 non-null  object 
 3   room_type               74111 non-null  object 
 4   amenities               74111 non-null  object 
 5   accommodates            74111 non-null  int64  
 9   cleaning_fee            74111 non-null  bool   
 10  city                    74111 non-null  object 
 24  review_scores_rating    57389 non-null  float64
 26  zipcode                 73145 non-null  object 
 27  bedrooms                74020 non-null  float64

df_airbnb['id'].duplicated().sum() # total de duplicados en la columna id

df_airbnb['bedrooms'].isna().sum()

df_airbnb[['accommodates', 'bedrooms']].head(20) # verificar cuál es el patrón que sigue el número de personas y número de habitaciones

df_airbnb['amenities'][0]

df_airbnb['review_scores_rating'].isna().sum() # total de valores nulos en la califacion

df_airbnb[df_airbnb['review_scores_rating'].isna()]['number_of_reviews'].value_counts() # conteo de propiedades con calificación nula y número de reviews

zero_count = (df_airbnb['accommodates'] == 0).sum()
print(f"Cantidad de valores cero en 'accommodates': {zero_count}") # verificar que no haya una propiedad que aloje 0 personas

df_airbnb['property_type'].value_counts() # conteo de tipos de propiedades

df_airbnb['room_type'].value_counts() # conteo de tipos de habitaciones

df_airbnb['amenities'].value_counts() # conteo de amenidades

df_airbnb['city'].value_counts() # conteo de ciudades
