==> notebooks/2_data_wrangling.ipynb <==
import pandas as pd
import requests
import time
import os

df_airbnb = pd.read_csv('/Users/melaniealvarez/Documents/Octavo semestres/Data Mining/Pset2/data/raw/Airbnb_Data.csv')

Completar los valores faltantes en zipcode con la información latitud, longitud y el API de Google Maps

API_KEY = "AIzaSyDtjjoF7nwKqOKNEtuxgXH97-ZsosbklE4"

def get_zipcode(lat, lon):
    """Obtiene el código postal usando Google Maps API a partir de latitud y longitud."""
    url = f"https://maps.googleapis.com/maps/api/geocode/json?latlng={lat},{lon}&key={API_KEY}"
    response = requests.get(url)
    data = response.json()
    
    zipcode = None
    
    if "results" in data and len(data["results"]) > 0:
        for result in data["results"]:
            for component in result["address_components"]:
                if "postal_code" in component["types"]:
                    zipcode = component["long_name"]
                    return zipcode  # Retorna en cuanto encuentra el código postal
    
    return zipcode

# Filtrar los valores NaN en zipcode
missing_zipcodes = df_airbnb[df_airbnb["zipcode"].isna()]

for index, row in missing_zipcodes.iterrows():
    lat, lon = row["latitude"], row["longitude"]
    zipcode = get_zipcode(lat, lon)
    
    if zipcode:
        df_airbnb.at[index, "zipcode"] = zipcode
    
    time.sleep(1)  # Para evitar superar el límite de peticiones de la API

print("Proceso completado: ZIP codes actualizados.")

Reemplazar aquellas propiedades con bedrooms = Nan o bedrooms = 0, con la mitad del número de accommodates

# Reemplazar bedrooms igual a 0 o NaN con la mitad del valor en accommodates
df_airbnb['bedrooms'] = df_airbnb.apply(lambda row: row['accommodates'] / 2 if pd.isna(row['bedrooms']) or row['bedrooms'] == 0 else row['bedrooms'], axis=1)

Reemplazar NaN en review_scores_rating por la media

df_airbnb['missing_rating'] = df_airbnb['review_scores_rating'].isna().astype(int)

# Reemplazar NaN con el promedio de la columna review_scores_rating
mean_rating = df_airbnb['review_scores_rating'].mean()
df_airbnb['review_scores_rating'].fillna(mean_rating, inplace=True)

df_airbnb.info() # Verificar que no haya valores NaN en las variables anteriores

Analizar outliers en la variable log_price

# Calcular los percentiles 25 y 75
Q1 = df_airbnb["log_price"].quantile(0.25)
Q3 = df_airbnb["log_price"].quantile(0.75)

# Calcular el rango intercuartil (IQR)
IQR = Q3 - Q1

# Definir los límites inferior y superior
lower_bound = Q1 - 1.5 * IQR
upper_bound = Q3 + 1.5 * IQR

outliers_iqr = df_airbnb[(df_airbnb["log_price"] < lower_bound) | (df_airbnb["log_price"] > upper_bound)]

# Mostrar los valores atípicos detectados
#print("Outliers detectados en la duración de las llamadas (IQR):")
#print(outliers_iqr)

# Obtener el valor más alto dentro de los outliers detectados
max_outlier = outliers_iqr["log_price"].max()
min_outlier = outliers_iqr["log_price"].min()

# Imprimir el valor más alto
print(f"El valor más alto entre los outliers es: {max_outlier}")
print(f"El valor más alto entre los outliers es: {min_outlier}")


lower_bound

upper_bound

folder_path = "/Users/melaniealvarez/Documents/Octavo semestres/Data Mining/Pset2/data/processed"
file_path = os.path.join(folder_path, "airbnb_clean.csv")
df_airbnb.to_csv(file_path, index=False)
