==> notebooks/3_feature_engineering.ipynb <==
import pandas as pd
import os

df_airbnb_clean = pd.read_csv('/Users/melaniealvarez/Documents/Octavo semestres/Data Mining/Pset2/data/processed/airbnb_clean.csv')

Verificar que las variables anteriormente manipuladas no tengan valores nulos

df_airbnb_clean['review_scores_rating'].isna().sum()

df_airbnb_clean['zipcode'].isna().sum()

df_airbnb_clean['bedrooms'].isna().sum()

Nuevo dataframe con las variables pertinentes para el modelo

# Lista de columnas a conservar
columns_to_keep = [
    'id', 'log_price', 'property_type', 'room_type', 'amenities',
    'accommodates', 'cleaning_fee', 'city', 'review_scores_rating',
    'bedrooms'
]

# Crear un nuevo DataFrame con solo estas columnas
df_airbnb_ml = df_airbnb_clean[columns_to_keep]

# Mostrar información del nuevo DataFrame
print(df_airbnb_ml.info())

One-Hot Encoding para room_type

df_airbnb_ml = pd.get_dummies(df_airbnb_ml, columns=['room_type'], prefix='room')

room_cols = [col for col in df_airbnb_ml.columns if col.startswith('room_')]
df_airbnb_ml[room_cols] = df_airbnb_ml[room_cols].astype(int)

Convertir valores bool true/false en 0 y 1

df_airbnb_ml = pd.get_dummies(df_airbnb_ml, columns=['city'], prefix='city')

room_cols_city = [col for col in df_airbnb_ml.columns if col.startswith('city_')]
df_airbnb_ml[room_cols_city] = df_airbnb_ml[room_cols_city].astype(int)

df_airbnb_ml['amenities'] = df_airbnb_ml['amenities'].apply(lambda x: 0 if x == '{}' else 1)

df_airbnb_ml['cleaning_fee'] = df_airbnb_ml['cleaning_fee'].astype(int)

from sklearn.preprocessing import LabelEncoder

Label Encoding a 'property_type'

le = LabelEncoder()

df_airbnb_ml['property_type_encoded'] = le.fit_transform(df_airbnb_ml['property_type'])

print(df_airbnb_ml[['property_type', 'property_type_encoded']].head())


df_airbnb_ml.info()

folder_path = "/Users/melaniealvarez/Documents/Octavo semestres/Data Mining/Pset2/data/ml"
file_path = os.path.join(folder_path, "airbnb_analysis.csv")
df_airbnb_ml.to_csv(file_path, index=False)
